+++ CHANGELOG.md
@@ -39,6 +39,7 @@ release.
 - Added residual column information to the Points model
 
 ### Changed
+- Roi `is_valid` method changed to use np.isclose to ensure floating point tolerances when checking if the NDV is in the ROI.
 - `geom_match_simple` defaults to a 3rd order warp for interpolation
 - Speed improvements for place_points_from_cnet dependent on COPY method instead of ORM update
 - License from custom to CC0. Fixes [#607](https://github.com/USGS-Astrogeology/autocnet/issues/607)